refactor pixel_ball
